remove unused block import channel & migrate txpool db port
