+++ crates/services/consensus_module/poa/Cargo.toml
@@ -12,7 +12,6 @@ version = "0.15.1"
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
-fuel-core-interfaces = { path = "../../../fuel-core-interfaces", version = "0.15.1" }
 fuel-core-storage = { path = "../../../storage", version = "0.15.1" }
 fuel-core-types = { path = "../../../types", version = "0.15.1" }
 humantime-serde = "1.1.1"
